% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par.uncertainty.thr.function.R
\name{par.uncertainty.thr}
\alias{par.uncertainty.thr}
\title{Title}
\usage{
par.uncertainty.thr(
  vario.mod.output,
  mod.nr,
  par.est = NULL,
  data = NULL,
  max.dist = NULL,
  nbins = NULL,
  B = 1000,
  threshold.factor = 1.2
)
}
\arguments{
\item{threshold.factor}{}
}
\value{

}
\description{
Title
}
